requirements.txt
awsebcli==3.12.1
boto==2.48.0
boto3==1.7.36
botocore==1.10.36
Django==2.0.1
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.4
django-storages==1.6.6
jsonschema==2.6.0
markdown2==2.3.5
MarkupSafe==1.0
mpmath==1.0.0
numpy==1.13.3
pandas==0.20.3
pandocfilters==1.4.2
pathspec==0.5.0
pickleshare==0.7.4
Pillow==5.0.0
product==0.1.1.1
prompt-toolkit==1.0.15
protobuf==3.1.0
psycopg2==2.7.1
pydatalog==0.17.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-mnist==0.3
pytz==2017.2
pyyaml==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.9.1
s3transfer==0.1.11
semantic-version==2.5.0
sequential==1.0.0
simplegeneric==0.8.1
six==1.11.0
sympy==1.1.1
tabulate==0.7.5
termcolor==1.1.0
testpath==0.3.1
transliterate==1.10
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
whitenoise==3.3.1
widgetsnbextension==3.0.3
dj_database_url==0.5.0
gunicorn==19.6.0
djangorestframework==3.8.1
xlrd==1.1.0
django_cors_headers==2.4.0